Redesigned filtering
Completely redesigned filtering. When we have over 500 hands on a player we filter most relevant hands to use for the stats. Before we added a filter like "hands with 5 players (like now)" and maybe ended up with 135 hands. Now instead of filtering, we rank each hand based on their relevance and pick exactly 500 most relevant hands.

Improved performance. We stop reading files once we find that hero got eliminated or won. We also stop reading files that don't contain any hands for the last hour. 

TODO: We just introduced a new bug: some hands are stored multiple times. Need to get back on that. The best solution might be to save the number of last line processed for each file. We can then skip all lines we've already read.